+++ src/operations.ts
@@ -14,7 +14,7 @@ export function setField(
     type: "op",
     operations: {
       ordered: [],
-      state: { [nodeId]: { [field]: JSON.stringify(value) } },
+      state: { [nodeId]: { [field]: value } },
     },
   };
 }